Move ServiceCall to api, make possible to create ServiceCall with fixed destination address with no routing.
