+++ services/src/main/java/io/scalecube/services/Microservices.java
@@ -13,6 +13,8 @@ import io.scalecube.services.methods.ServiceMethodRegistryImpl;
 import io.scalecube.services.metrics.Metrics;
 import io.scalecube.services.registry.ServiceRegistryImpl;
 import io.scalecube.services.registry.api.ServiceRegistry;
+import io.scalecube.services.routing.RoundRobinServiceRouter;
+import io.scalecube.services.routing.Routers;
 import io.scalecube.services.transport.api.ClientTransport;
 import io.scalecube.services.transport.api.DataCodec;
 import io.scalecube.services.transport.api.ServerTransport;
@@ -187,7 +189,8 @@ public class Microservices {
 
               // invoke service providers and register services
               List<Object> serviceInstances =
-                  serviceProviders.stream()
+                  serviceProviders
+                      .stream()
                       .flatMap(serviceProvider -> serviceProvider.provide(call).stream())
                       .peek(this::registerInMethodRegistry)
                       .peek(
@@ -201,7 +204,9 @@ public class Microservices {
                   .create(serviceEndpointBuilder.build(), serviceRegistry)
                   .publishOn(scheduler)
                   .then(Mono.defer(() -> startGateway(call)).publishOn(scheduler))
-                  .then(Mono.fromCallable(() -> Reflect.inject(this, serviceInstances)))
+                  .then(
+                      Mono.fromCallable(
+                          () -> MicroservicesInjection.inject(this, serviceInstances)))
                   .then(Mono.fromCallable(() -> JmxMonitorMBean.start(this)))
                   .then(Mono.defer(discoveryBootstrap::start).publishOn(scheduler))
                   .thenReturn(this);
@@ -235,7 +240,8 @@ public class Microservices {
   }
 
   public ServiceCall call() {
-    return new ServiceCall(transportBootstrap.clientTransport, serviceRegistry, methodRegistry);
+    return new ServiceCall(transportBootstrap.clientTransport, serviceRegistry, methodRegistry)
+        .router(Routers.getRouter(RoundRobinServiceRouter.class));
   }
 
   public List<Gateway> gateways() {
@@ -529,7 +535,8 @@ public class Microservices {
     }
 
     private Gateway gateway(String id) {
-      return gateways.stream()
+      return gateways
+          .stream()
           .filter(gw -> gw.id().equals(id))
           .findFirst()
           .orElseThrow(
@@ -726,7 +733,9 @@ public class Microservices {
 
     @Override
     public Collection<String> getGatewayAddresses() {
-      return microservices.gateways().stream()
+      return microservices
+          .gateways()
+          .stream()
           .map(gw -> gw.id() + " -> " + gw.address())
           .collect(Collectors.toList());
     }
@@ -745,7 +754,10 @@ public class Microservices {
 
     @Override
     public Collection<String> getServiceEndpoints() {
-      return microservices.serviceRegistry.listServiceEndpoints().stream()
+      return microservices
+          .serviceRegistry
+          .listServiceEndpoints()
+          .stream()
           .map(ServiceEndpoint::toString)
           .collect(Collectors.toList());
     }
