+++ services/src/main/java/io/scalecube/services/methods/ServiceMethodRegistryImpl.java
@@ -3,7 +3,6 @@ package io.scalecube.services.methods;
 import io.scalecube.services.Reflect;
 import io.scalecube.services.exceptions.ServiceProviderErrorMapper;
 import io.scalecube.services.transport.api.ServiceMessageDataDecoder;
-
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
